Update Sector Size calculation
Currently it just returns the total length of all sides of the polygon that makes up the sector.  It really should return the actual area of that polygon.  The second double it returns will remain the same, the average length of the sides of the polygon, at least until a proper utility method/library is created.